+++ internal/controller/compute/cubeserver/cubeserver.go
@@ -170,8 +170,7 @@ func (c *externalServer) Create(ctx context.Context, mg resource.Managed) (manag
 		// Servers should have unique names per datacenter.
 		// Check if there are any existing servers with the same name.
 		// If there are multiple, an error will be returned.
-		instance, err := c.service.CheckDuplicateCubeServer(ctx, cr.Spec.ForProvider.DatacenterCfg.DatacenterID,
-			cr.Spec.ForProvider.Name, templateID, cr.Spec.ForProvider.CPUFamily)
+		instance, err := c.service.CheckDuplicateCubeServer(ctx, cr.Spec.ForProvider.DatacenterCfg.DatacenterID, cr.Spec.ForProvider.Name, templateID)
 		if err != nil {
 			return managed.ExternalCreation{}, err
 		}